backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
tensorflow==2.15.0
numpy==1.26.2
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78 